requirements.txt
fastapi
uvicorn[standard]
torch
diffusers
transformers
accelerate
minio
python-dotenv
sentencepiece
imageio
imageio-ffmpeg
